Also known as: Don't rebuild OverlayEntries when one changes opaqueness.

## Description

Prior to this change we were re-arranging the tree when an OverlayEntry went from onstage to offstage (due to opaqueness changes) and vice versa:

https://github.com/flutter/flutter/blob/4f9b6cf01799a60ec6c4cc55c614a0e76ca80a97/packages/flutter/lib/src/widgets/overlay.dart#L461-L467

Since the OverlayEntries have a global key their corresponding element was deactivated at the old location inside the Stack and re-activated at the new location outside the Stack. This caused all StatefulWidgets within that subtree to rebuild because their ancestor had changed.

This change avoids the rebuild by not re-arranging the tree. Instead, both onstage and offstage OverlayEntries are children of the _Theatre, so they don't have to be moved around in the tree. The RenderObject of the _Theatre implements a special version of a Stack that just ignores the offstage children during rendering and layout. (Since the offstage children are always all at the beginning of the child list, this is implemented by providing an integer skipCount indicating the number of children to skip.)

The unnecessary rebuilds were most commonly felt when pushing new Routes because Navigator uses an Overlay internally. 

## Related Issues

Fixes https://github.com/flutter/flutter/issues/45797.

## Tests

I added the following tests:

* OverlayEntries do not rebuild when opaque entry is added
* OverlayEntries do not rebuild when opaqueness changes
* entries below opaque entries are ignored for hit testing
* Semantics of entries below opaque entries are ignored
* Can used Positioned within OverlayEntry
* Pushing opaque Route does not rebuild routes below

## Checklist

Before you create this PR confirm that it meets all requirements listed below by checking the relevant checkboxes (`[x]`). This will ensure a smooth and quick review process.

- [x] I read the [Contributor Guide] and followed the process outlined there for submitting PRs.
- [x] I signed the [CLA].
- [x] I read and followed the [Flutter Style Guide], including [Features we expect every widget to implement].
- [x] I read the [Tree Hygiene] wiki page, which explains my responsibilities.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] All existing and new tests are passing.
- [x] The analyzer (`flutter analyze --flutter-repo`) does not report any problems on my PR.
- [x] I am willing to follow-up on review comments in a timely manner.

## Breaking Change

Did any tests fail when you ran them? Please read [Handling breaking changes].

- [ ] No, no existing tests failed, so this is *not* a breaking change.
- [ ] Yes, this is a breaking change. *If not, delete the remainder of this section.*
   - [ ] I wrote a design doc: https://flutter.dev/go/template *Replace this with a link to your design doc's short link*
   - [ ] I got input from the developer relations team, specifically from: *Replace with the names of who gave advice*
   - [ ] I wrote a migration guide: *Replace with a link to your migration guide*

<!-- Links -->
[issue database]: https://github.com/flutter/flutter/issues
[Contributor Guide]: https://github.com/flutter/flutter/wiki/Tree-hygiene#overview
[Test Coverage]: https://github.com/flutter/flutter/wiki/Test-coverage-for-package%3Aflutter
[Flutter Style Guide]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo
[Features we expect every widget to implement]: https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo#features-we-expect-every-widget-to-implement
[CLA]: https://cla.developers.google.com/
[Tree Hygiene]: https://github.com/flutter/flutter/wiki/Tree-hygiene
[Handling breaking changes]: https://github.com/flutter/flutter/wiki/Tree-hygiene#handling-breaking-changes