@goderbauer 2020-01-15
@Hixie Even though I had to update framework tests for this change I would like to handle it as a non-breaking change. Can you grant an exception? 

The changes to the tests were to update error messages and error-message-like output. In one instance I had to change a type due to the changed internals of _Theatre.