@goderbauer 2020-01-15
> I may be missing something, but the skipCount seems sketchy to me. I think we shoudl at least be asserting that the children we think are offstage are in fact offstage, and vice-versa.

The definition of the (private) _Theatre widget is: "give me a list of children and a number n. I'll not layout/render the first n children (effectively putting them offstage)."

The number n defines what is going to be offstage. There's no other mechanism to tell of a child should be offstage or not...