@Hixie 2020-01-21
I haven't reviewed the test of the patch (I'm on my phone) but the test changes call be considered exempt from the breaking change policy as they are only testing internals.