@goderbauer 2020-01-22
This change positively affected some of our Gallery transition benchmarks:

<img width="270" alt="Screen Shot 2020-01-22 at 9 24 29 AM" src="https://user-images.githubusercontent.com/1227763/72917615-0369a580-3cf9-11ea-8401-d8ae70e11a81.png">

* dropped from 63 frames to 50 (minus 21%)

<img width="279" alt="Screen Shot 2020-01-22 at 9 24 23 AM" src="https://user-images.githubusercontent.com/1227763/72917616-0369a580-3cf9-11ea-9e72-000b0af32435.png">

* dropped from 50.45ms to 36.64ms (minus 27%)

<img width="297" alt="Screen Shot 2020-01-22 at 9 21 54 AM" src="https://user-images.githubusercontent.com/1227763/72917429-acfc6700-3cf8-11ea-8e58-ddf53c09dac8.png">

* dropped from 7.38ms to 6.13ms (minus 17%)