@dnfield 2020-01-22
Wow I'm so used to seeing posts like that linked to a today bug due to a regression.  Nice!