@dkwingsmt 2020-01-23
If we were to reland this PR, make sure to address the issue mentioned in https://github.com/flutter/flutter/pull/49318, where the order between multiple dependencies are not consistent. They might change due to an irrelevant change, or be different between platforms.